feat: car class 추가 test: car 생성 테스트 추가
